+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2020-4-30]
+### Issues
+
+- No audio when firefox captureStream form audio element.
+
+
 ## [2020-4-29]
 ### Added
 - 使用FileReader读取上传音频文件数据，使用 decodeAudioData 异步解码音频文件中的 ArrayBuffer。最后生成stream，再转换为ogg.